warning
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@7778 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/bbs2www/html/bbssnd.php
@@ -32,10 +32,13 @@ login_init();
 		} else {
 			$tmpl = 1;
 			$filename = "tmp/".$currentuser["userid"].".tmpl.tmp";
-			$handle = fopen($filename, "r");
-			$contents = fread($handle, filesize($filename));
-			fclose($handle);
-			unlink($filename);
+			if (($handle = @fopen($filename, "r")) !== false) {
+				$contents = fread($handle, filesize($filename));
+				fclose($handle);
+				unlink($filename);
+			} else {
+				html_error_quit("无法读取文件");
+			}
 		}
 		if (isset($_GET["reid"])) $reID = $_GET["reid"];
 		else
